requirements.txt
fritzconnection==1.5.0
requests==2.25.1
PyYAML==5.4.1
websocket-client~=0.57.0